requirements.txt
Django==2.2.7
djangorestframework==3.11.0
black
django-filter
bs4
requests